+++ src/Optional/Extensions/AsOptionExtensions.cs
@@ -11,6 +11,6 @@ public static class AsOptionExtensions
     /// <summary>
     /// Converts a nullable object to an <see cref="Option{TValue}"/>.
     /// </summary>
-    public static async Task<Option<TValue>> AsOption<TValue>(this Task<TValue?> task) =>
-        AsOption(await task.ConfigureAwait(false));
+    public static async Task<Option<TValue>> AsOption<TValue>(this Task<TValue?> task) => 
+        (await task.ConfigureAwait(false)).AsOption();
 }